ia64: fix byteswap conversion warning
Same fix as went in for other architectures.

URL: http://sourceware.org/bugzilla/show_bug.cgi?id=12194

+++ ports/sysdeps/ia64/bits/byteswap.h
@@ -25,7 +25,7 @@
 
 /* Swap bytes in 16 bit value.  */
 #define __bswap_constant_16(x) \
-     ((((x) >> 8) & 0xff) | (((x) & 0xff) << 8))
+     ((unsigned short int)((((x) >> 8) & 0xff) | (((x) & 0xff) << 8)))
 
 /* Get __bswap_16.  */
 #include <bits/byteswap-16.h>
